D3D Render target set

+++ Code/Engine/Core/StringUtils.hpp
@@ -9,4 +9,4 @@ const std::string Stringf( const int maxLength, const char* format, ... );
 /// assume \0 at end of string
 std::vector<std::string> Split(const char* cstring, char delimiter = ' ', bool reserveDelimiter = false);
 
-int LoadTextFileToString(const std::string& filePath, std::string& out_string);
+size_t LoadTextFileToString(const std::string& filePath, std::string& out_string);
